backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0

django-cors-headers
djangorestframework-simplejwt
Pillow
python-dotenv
django-autoslug
whitenoise
drf-yasg
redis
django-ratelimit
django-filter